crossdock - Start moving crossdock client functionality into a lib, starting with combinations
